requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
neo4j==5.23.0
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.8.2

